+++ src/main/java/jeje/work/aeatbe/controller/KakaoAuthController.java
@@ -5,6 +5,7 @@ import java.io.IOException;
 import jeje.work.aeatbe.domian.KakaoProperties;
 import jeje.work.aeatbe.domian.KakaoTokenResponsed;
 import jeje.work.aeatbe.dto.user.TokenResponseDto;
+import jeje.work.aeatbe.service.KakaoService;
 import jeje.work.aeatbe.service.UserService;
 import jeje.work.aeatbe.utility.JwtUtil;
 import lombok.RequiredArgsConstructor;
@@ -18,11 +19,12 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequestMapping("/api/users")
 @RequiredArgsConstructor
-public class UserController {
+public class KakaoAuthController {
 
     private final UserService userService;
     private final JwtUtil jwtUtil;
     private final KakaoProperties kakaoProperties;
+    private final KakaoService kakaoService;
 
     @GetMapping("/login")
     public void redirectKakaoLogin(HttpServletResponse response) throws IOException {
@@ -34,8 +36,8 @@ public class UserController {
 
     @GetMapping("/callback")
     public ResponseEntity<TokenResponseDto> getAccessToken(@RequestParam String code){
-        KakaoTokenResponsed token = userService.getKakaoTokenResponse(code);
-        String jwt = userService.Login(token.accessToken(), token.refreshToken());
+        KakaoTokenResponsed token = kakaoService.getKakaoTokenResponse(code);
+        String jwt = kakaoService.Login(token.accessToken(), token.refreshToken());
         return new ResponseEntity<>(new TokenResponseDto(jwt), HttpStatus.OK);
     }
