+++ src/main/java/jeje/work/aeatbe/service/UserService.java
@@ -17,22 +17,14 @@ import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.web.client.RestClient;
 
 @Service
+@RequiredArgsConstructor
 public class UserService {
 
     private final UserRepository userRepository;
     private final JwtUtil jwtUtil;
     private final ObjectMapper objectMapper;
-    private final KakaoProperties kakaoProperties;
-    private final RestClient restClient;
+    private final KakaoService kakaoService;
 
-    public UserService(UserRepository userRepository, JwtUtil jwtUtil,
-        ObjectMapper objectMapper, KakaoProperties kakaoProperties) {
-        this.userRepository = userRepository;
-        this.jwtUtil = jwtUtil;
-        this.objectMapper = objectMapper;
-        this.kakaoProperties = kakaoProperties;
-        this.restClient = RestClient.builder().build();
-    }
 
 
     public Long getUserId(String kakaoId){
@@ -52,47 +44,8 @@ public class UserService {
         return userRepository.findByKakaoId(kakaoId).isPresent();
     }
 
-    public KakaoTokenResponsed getKakaoTokenResponse(String code){
-        var uri = "https://kauth.kakao.com/oauth/token";
-        var body = createBody(code);
-        var response = restClient.post()
-            .uri(URI.create(uri))
-            .contentType(MediaType.APPLICATION_FORM_URLENCODED)
-            .body(body)
-            .retrieve()
-            .body(KakaoTokenResponsed.class);
-        return response;
-    }
 
-    private LinkedMultiValueMap<String, String> createBody(String code) {
-        var body = new LinkedMultiValueMap<String, String>();
-        body.add("grant_type", "authorization_code");
-        body.add("client_id", kakaoProperties.clientId());
-        body.add("redirect_uri", kakaoProperties.redirectUrl());
-        body.add("code", code);
-        return body;
-    }
 
-    @Transactional
-    public String Login(String accessToken, String refreshToken){
-        var uri = "https://kapi.kakao.com/v2/user/me";
-        var response = restClient.get()
-            .uri(URI.create(uri))
-            .header("Authorization", "Bearer " + accessToken)
-            .retrieve()
-            .body(KakaoUserInfo.class);
-        String kakaoId = response.id()+"";
-        Optional<User> user = userRepository.findByKakaoId(kakaoId);
-        if(user.isEmpty()){
-            User newUser = User.builder().kakaoId(kakaoId).
-                accessToken(accessToken).
-                refreshToken(refreshToken).
-                build();
-             userRepository.save(newUser);
-             return jwtUtil.createToken(newUser);
-        }
-        user.get().kakaoTokenUpdate(accessToken, refreshToken);
-        return jwtUtil.createToken(user.get());
-    }
+
 
 }
